Added test functions for improved weather.map_meta to test_weather.py instead of a standalone file for better organization
